backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.3
dnspython==2.8.0
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
idna==3.13
Mako==1.3.12
MarkupSafe==3.0.3
motor==3.7.1
passlib==1.7.4
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
pymongo==4.17.0
python-dotenv==1.2.2
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
python-jose[cryptography]==3.3.0
pydantic[email]==2.13.3
python-multipart==0.0.20
bcrypt==4.0.1